move interfaces to interfaces.ts
